add routes for inernal gallery
